lib: change type of EID to u32 other than usize
In chapter 3, RISC-V SBI specification, it quotes:

```
In the name of compatibility, SBI extension IDs (EIDs) and SBI function IDs (FIDs) are encoded as signed 32-bit integers.
```

We mark extension ID to 32-bit to reflect such definition.

+++ src/base.rs
@@ -1,7 +1,7 @@
 //! Chapter 4. Base Extension (EID #0x10).
 
 /// Extension ID for RISC-V SBI Base extension.
-pub const EID_BASE: usize = 0x10;
+pub const EID_BASE: u32 = 0x10;
 pub use fid::*;
 
 /// Default probe value for the target SBI extension is unavailable.
